Refactor: moving ChirpCore part 2, namespace updates Co-authored-by: Maria <arnv@itu.dk> Co-authored-by: Henriette <hoes@itu.dk> Co-authored-by: Laura <llsa@itu.dk>
