+++ src/ChirpCore/ChirpDBContext.cs
@@ -2,11 +2,12 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using ChirpCore.Domain;
 
-namespace Chirp.EFCore;
+namespace ChirpCore;
 
 public class ChirpDBContext : IdentityDbContext<Author>
 {
-	public ChirpDBContext(DbContextOptions<ChirpDBContext> options) : base(options) {
+	public ChirpDBContext(DbContextOptions<ChirpDBContext> options) : base(options)
+	{
 
 	}
 	public ChirpDBContext()
@@ -29,7 +30,7 @@ public class ChirpDBContext : IdentityDbContext<Author>
 								.IsUnique();
 		/*modelBuilder.Entity<Author>()
 								.HasKey(k => new { k.FollowerId, k.FollowingId });*/
-								//above will be relevant later
+		//above will be relevant later
 	}
 	public DbSet<Cheep> Cheeps { get; set; }
