--- src/ChirpCore/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace ChirpCore;
-
-public class Class1
-{
-
-}
