+++ src/ChirpCore/DTOs/AuthorDTO.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace ChirpCore.DTOs
 {
-    public class AuthorDTO
-    {
-		public int UserId { get; set; }
+	public class AuthorDTO
+	{
+		public int AuthorId { get; set; }
 		public string? Name { get; set; }
 		//public string Email { get; set; }
 		//public ICollection<Cheep> Cheeps { get; set; }
