+++ src/ChirpCore/DTOs/CheepDTO.cs
@@ -9,9 +9,9 @@ namespace ChirpCore.DTOs
 	public class CheepDTO
 	{
 		public int AuthorID { get; set; }
-		[StringLength(160)] public string Text { get; set; }
+		[StringLength(160)] public string? Text { get; set; }
 		public DateTime TimeStamp { get; set; }
-		public string AuthorName { get; set; }
+		public string? AuthorName { get; set; }
 
 	}
 }
