+++ src/ChirpCore/Domain/Author.cs
@@ -4,12 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 
-
 namespace ChirpCore.Domain
 {
-    public class Author : IdentityUser
-    {
-		public int UserId { get; set; }
+	public class Author : IdentityUser
+	{
+		public int AuthorId { get; set; }
 		public required string Name { get; set; }
 		//public required string? Email { get; set; }
 		public required ICollection<Cheep> Cheeps { get; set; }
