+++ src/Program.cs
@@ -1,4 +1,3 @@
-using Chirp.EFCore;
 using ChirpCore;
 using ChirpCore.Domain;
 using ChirpCore.DTOs;
@@ -12,12 +11,12 @@ using Microsoft.Extensions.DependencyInjection;
 
 
 if (!Directory.Exists("/tmp/data"))
-    {
-            string DBFilePath = Path.GetTempPath();
-            string DBFilePathWithFile = Path.Combine(DBFilePath + "chirp.db");
-            Directory.CreateDirectory(DBFilePath);
-            File.Create(DBFilePathWithFile);
-    }
+{
+    string DBFilePath = Path.GetTempPath();
+    string DBFilePathWithFile = Path.Combine(DBFilePath + "chirp.db");
+    Directory.CreateDirectory(DBFilePath);
+    File.Create(DBFilePathWithFile);
+}
 
 
 var builder = WebApplication.CreateBuilder(args);
@@ -43,9 +42,9 @@ var app = builder.Build();
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
-	app.UseExceptionHandler("/Error");
-	// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-	app.UseHsts();
+    app.UseExceptionHandler("/Error");
+    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    app.UseHsts();
 }
 app.UseHttpsRedirection();
 
@@ -63,10 +62,10 @@ app.MapRazorPages();
 //Below 'using' block from Group 3. Seeds our database, and ensures that the database is created
 using (var scope = app.Services.CreateScope())
 {
-	var services = scope.ServiceProvider;
-	var context = services.GetRequiredService<ChirpDBContext>();
-	context.Database.EnsureCreated();
-	DbInitializer.SeedDatabase(context);
+    var services = scope.ServiceProvider;
+    var context = services.GetRequiredService<ChirpDBContext>();
+    context.Database.EnsureCreated();
+    DbInitializer.SeedDatabase(context);
 }
 
 app.Run();
